webapp/requirements.txt
Flask==3.0.0
Flask-Cors==5.0.0
Pillow==11.1.0
qrcode==8.0
pyzbar==0.1.9
gunicorn==21.2.0
pycryptodome==3.19.1
numpy==1.26.3
opencv-python-headless==4.9.0.80
